use fake encoder stepcounter on EL70?1
newer EL7041 have a dedicated CoE parameter for the stepcounter, which can
be used even if we use the internal encoder. However this means selecting
a non-standard PDO configuration, which does not work on older EL7041.

As we do not use the encoder of the EL7041, we can simply use its counter
value, which is the stepcounter if no encoder is present.